requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
databricks-connect==16.1.1
databricks-sdk==0.57.0
Flask==3.1.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
google-auth==2.38.0
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
propcache==0.3.0
protobuf==5.29.3
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
requests==2.32.3
rsa==4.9
setuptools==76.0.0
simple-websocket==1.1.0
six==1.17.0
slack_bolt==1.22.0
slack_sdk==3.34.0
sqlglot==26.10.1
tzdata==2025.1
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
python-dotenv==1.0.0